# 8304932: MethodTypeDescImpl can be mutated by argument passed to MethodTypeDesc.of