--- a/test/jdk/java/lang/constant/MethodTypeDescTest.java
+++ b/test/jdk/java/lang/constant/MethodTypeDescTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, 2019, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2018, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -24,6 +24,7 @@
 import java.lang.invoke.MethodType;
 import java.lang.constant.ClassDesc;
 import java.lang.constant.MethodTypeDesc;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.stream.IntStream;
@@ -31,15 +32,14 @@ import java.util.stream.Stream;
 
 import org.testng.annotations.Test;
 
-import static java.lang.constant.ConstantDescs.CD_int;
-import static java.lang.constant.ConstantDescs.CD_void;
+import static java.lang.constant.ConstantDescs.*;
 import static java.util.stream.Collectors.joining;
 import static java.util.stream.Collectors.toList;
-import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.fail;
+import static org.testng.Assert.*;
 
 /**
  * @test
+ * @bug 8304932
  * @compile MethodTypeDescTest.java
  * @run testng MethodTypeDescTest
  * @summary unit tests for java.lang.constant.MethodTypeDesc
@@ -104,12 +104,9 @@ public class MethodTypeDescTest extends SymbolicDescTest {
         }
 
         // try with null parameter
-        try {
+        expectThrows(NullPointerException.class, () -> {
             MethodTypeDesc newDesc = mtDesc.changeReturnType(null);
-            fail("should fail with NPE");
-        } catch (NullPointerException ex) {
-            // good
-        }
+        });
 
         // changeParamType
         for (int i=0; i<paramTypes.length; i++) {
@@ -133,6 +130,15 @@ public class MethodTypeDescTest extends SymbolicDescTest {
             MethodTypeDesc newDesc = mtDesc.dropParameterTypes(i, i + 1);
             assertEquals(newDesc, MethodTypeDesc.of(returnType, ps));
             testMethodTypeDesc(newDesc, mt.dropParameterTypes(i, i+1));
+
+            // drop multiple params
+            for (int j = i; j < paramTypes.length; j++) {
+                var t = new ArrayList<>(Arrays.asList(paramTypes));
+                t.subList(i, j).clear();
+                MethodTypeDesc multiDrop = mtDesc.dropParameterTypes(i, j);
+                assertEquals(multiDrop, MethodTypeDesc.of(returnType, t.toArray(ClassDesc[]::new)));
+                testMethodTypeDesc(multiDrop, mt.dropParameterTypes(i, j));
+            }
         }
 
         badDropParametersTypes(CD_void, paramDescs);
@@ -148,6 +154,21 @@ public class MethodTypeDescTest extends SymbolicDescTest {
                 assertEquals(newDesc, MethodTypeDesc.of(returnType, ps));
                 testMethodTypeDesc(newDesc, mt.insertParameterTypes(i, (Class<?>)p.resolveConstantDesc(LOOKUP)));
             }
+
+            // add multiple params
+            ClassDesc[] addition = {CD_int, CD_String};
+            var a = new ArrayList<>(Arrays.asList(paramTypes));
+            a.addAll(i, Arrays.asList(addition));
+
+            MethodTypeDesc newDesc = mtDesc.insertParameterTypes(i, addition);
+            assertEquals(newDesc, MethodTypeDesc.of(returnType, a.toArray(ClassDesc[]::new)));
+            testMethodTypeDesc(newDesc, mt.insertParameterTypes(i, Arrays.stream(addition).map(d -> {
+                try {
+                    return (Class<?>) d.resolveConstantDesc(LOOKUP);
+                } catch (ReflectiveOperationException ex) {
+                    throw new RuntimeException(ex);
+                }
+            }).toArray(Class[]::new)));
         }
 
         badInsertParametersTypes(CD_void, paramDescs);
@@ -158,47 +179,32 @@ public class MethodTypeDescTest extends SymbolicDescTest {
                 IntStream.rangeClosed(0, paramDescTypes.length - 1)
                         .mapToObj(i -> ClassDesc.ofDescriptor(paramDescTypes[i])).toArray(ClassDesc[]::new);
         MethodTypeDesc mtDesc = MethodTypeDesc.of(returnType, paramTypes);
-        try {
+        expectThrows(IndexOutOfBoundsException.class, () -> {
             MethodTypeDesc newDesc = mtDesc.insertParameterTypes(-1, paramTypes);
-            fail("pos < 0 should have failed");
-        } catch (IndexOutOfBoundsException ex) {
-            // good
-        }
+        });
 
-        try {
+        expectThrows(IndexOutOfBoundsException.class, () -> {
             MethodTypeDesc newDesc = mtDesc.insertParameterTypes(paramTypes.length + 1, paramTypes);
-            fail("pos > current arguments length should have failed");
-        } catch (IndexOutOfBoundsException ex) {
-            // good
-        }
+        });
 
-        try {
+        expectThrows(IllegalArgumentException.class, () -> {
             ClassDesc[] newParamTypes = new ClassDesc[1];
             newParamTypes[0] = CD_void;
             MethodTypeDesc newDesc = MethodTypeDesc.of(returnType, CD_int);
             newDesc = newDesc.insertParameterTypes(0, newParamTypes);
-            fail("shouldn't allow parameters with class descriptor CD_void");
-        } catch (IllegalArgumentException ex) {
-            // good
-        }
+        });
 
-        try {
+        expectThrows(NullPointerException.class, () -> {
             MethodTypeDesc newDesc = MethodTypeDesc.of(returnType, CD_int);
             newDesc = newDesc.insertParameterTypes(0, null);
-            fail("should fail with NPE");
-        } catch (NullPointerException ex) {
-            // good
-        }
+        });
 
-        try {
+        expectThrows(NullPointerException.class, () -> {
             ClassDesc[] newParamTypes = new ClassDesc[1];
             newParamTypes[0] = null;
             MethodTypeDesc newDesc = MethodTypeDesc.of(returnType, CD_int);
             newDesc = newDesc.insertParameterTypes(0, newParamTypes);
-            fail("should fail with NPE");
-        } catch (NullPointerException ex) {
-            // good
-        }
+        });
     }
 
     private void badDropParametersTypes(ClassDesc returnType, String... paramDescTypes) {
@@ -206,40 +212,26 @@ public class MethodTypeDescTest extends SymbolicDescTest {
                 IntStream.rangeClosed(0, paramDescTypes.length - 1)
                         .mapToObj(i -> ClassDesc.ofDescriptor(paramDescTypes[i])).toArray(ClassDesc[]::new);
         MethodTypeDesc mtDesc = MethodTypeDesc.of(returnType, paramTypes);
-        try {
+
+        expectThrows(IndexOutOfBoundsException.class, () -> {
             MethodTypeDesc newDesc = mtDesc.dropParameterTypes(-1, 0);
-            fail("start index < 0 should have failed");
-        } catch (IndexOutOfBoundsException ex) {
-            // good
-        }
+        });
 
-        try {
+        expectThrows(IndexOutOfBoundsException.class, () -> {
             MethodTypeDesc newDesc = mtDesc.dropParameterTypes(paramTypes.length, 0);
-            fail("start index = arguments.length should have failed");
-        } catch (IndexOutOfBoundsException ex) {
-            // good
-        }
+        });
 
-        try {
+        expectThrows(IndexOutOfBoundsException.class, () -> {
             MethodTypeDesc newDesc = mtDesc.dropParameterTypes(paramTypes.length + 1, 0);
-            fail("start index > arguments.length should have failed");
-        } catch (IndexOutOfBoundsException ex) {
-            // good
-        }
+        });
 
-        try {
+        expectThrows(IndexOutOfBoundsException.class, () -> {
             MethodTypeDesc newDesc = mtDesc.dropParameterTypes(0, paramTypes.length + 1);
-            fail("end index > arguments.length should have failed");
-        } catch (IndexOutOfBoundsException ex) {
-            // good
-        }
+        });
 
-        try {
+        expectThrows(IndexOutOfBoundsException.class, () -> {
             MethodTypeDesc newDesc = mtDesc.dropParameterTypes(1, 0);
-            fail("start index > end index should have failed");
-        } catch (IndexOutOfBoundsException ex) {
-            // good
-        }
+        });
     }
 
     public void testMethodTypeDesc() throws ReflectiveOperationException {
@@ -260,49 +252,41 @@ public class MethodTypeDescTest extends SymbolicDescTest {
                                               "(Ljava.lang.String;)V", "(java/lang/String)V");
 
         for (String d : badDescriptors) {
-            try {
+            expectThrows(IllegalArgumentException.class, () -> {
                 MethodTypeDesc r = MethodTypeDesc.ofDescriptor(d);
-                fail(d);
-            }
-            catch (IllegalArgumentException e) {
-                // good
-            }
+            });
         }
 
         // try with null argument
-        try {
+        expectThrows(NullPointerException.class, () -> {
             MethodTypeDesc r = MethodTypeDesc.ofDescriptor(null);
-            fail("should fail with NPE");
-        } catch (NullPointerException ex) {
-            // good
-        }
+        });
 
         // try with void arguments, this will stress another code path in particular
         // ConstantMethodTypeDesc::init
-        try {
+        expectThrows(IllegalArgumentException.class, () -> {
             MethodTypeDesc r = MethodTypeDesc.of(CD_int, CD_void);
-            fail("can't reach here");
-        }
-        catch (IllegalArgumentException e) {
-            // good
-        }
+        });
 
-        try {
+        expectThrows(NullPointerException.class, () -> {
             MethodTypeDesc r = MethodTypeDesc.of(CD_int, null);
-            fail("ClassDesc array should not be null");
-        }
-        catch (NullPointerException e) {
-            // good
-        }
+        });
 
-        try {
+        expectThrows(NullPointerException.class, () -> {
             ClassDesc[] paramDescs = new ClassDesc[1];
             paramDescs[0] = null;
             MethodTypeDesc r = MethodTypeDesc.of(CD_int, paramDescs);
-            fail("ClassDesc should not be null");
-        }
-        catch (NullPointerException e) {
-            // good
-        }
+        });
+    }
+
+    public void testImmutability() {
+        ClassDesc[] args = {CD_Object, CD_int};
+        var mtd = MethodTypeDesc.of(CD_void, args);
+
+        args[1] = CD_void;
+        assertEquals(mtd, MethodTypeDesc.of(CD_void, CD_Object, CD_int));
+
+        mtd.parameterArray()[1] = CD_void;
+        assertEquals(mtd, MethodTypeDesc.of(CD_void, CD_Object, CD_int));
     }
 }