@liach 2023-03-26
Additional comments:
1. Maybe it's feasible to add another factory:
```java
static MethodTypeDesc of(ClassDesc returnType, List<ClassDesc> parameterTypes) {
    return new MethodTypeDescImpl(returnType, List.copyOf(parameterTypes));
}
```
which would avoid another copy if `parameterTypes` is already a list from `List.of` factories.