@asotona 2023-04-17
> Additional comments:
> 
> 1. Maybe it's feasible to add another factory:
> 
> ```java
> static MethodTypeDesc of(ClassDesc returnType, List<ClassDesc> parameterTypes) {
>     return new MethodTypeDescImpl(returnType, List.copyOf(parameterTypes));
> }
> ```
> 
> which would avoid another copy if `parameterTypes` is already a list from `List.of` factories.
> 
> 2. Avoiding copies in `parameterList` can speed up its users, such as stack map generation in the Classfile API.

Yes, any improvements avoiding unnecessary arrays copying are welcome. `ClassDesc` and `MethodTypeDesc` are key pieces of the Classfile API with significant performance impact.