@mlchung 2023-04-17
Have you considered that the caller of instantiating `MethodTypeDescImpl` is responsible for passing a trusted array?  I think that `MethodTypeDescImpl` implementation already assumes it's a trusted array.  So `MethodTypeDesc::of` to call `new MethodTypeDescImpl` with a cloned array.   `MethodTypeDescImpl`  should call `new MethodTypeDescImpl`  instead of  `MethodTypeDesc::of` as this PR did. 

That way will avoid using `JavaUtilCollectionAccess` to create an immutable list without copying the input array.