@liach 2023-04-26
/issue JDK-8304932

Moved this patch to another issue that focus on converting the storage from array to immutable list. The fix for mutation is now a side effect of this patch.